requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0
asyncpg>=0.29
SQLAlchemy>=2.0
pydantic-settings>=2.2
alembic>=1.13
httpx[http2]>=0.27
apscheduler>=3.10
pyyaml>=6.0
playwright>=1.40
pytest-asyncio>=0.23
pytest>=8.2
aiosqlite>=0.19 